+++ src/Embed/Media.php
@@ -17,11 +17,22 @@ interface Media extends Markup
     /**
      * @return static
      */
-    public static function parse(string $embed): static;
+    public static function parse(
+        string $embed
+    ): static;
 
-    public static function extractProviderName(string $url): ?string;
-    public static function getClassForUrl(string $url): string;
-    public static function defaultUrlFromId(string $id): string;
+    public static function extractProviderName(
+        string $url
+    ): ?string;
+
+    public static function getClassForUrl(
+        string $url
+    ): string;
+
+    public static function defaultUrlFromId(
+        string $id
+    ): string;
+
 
     public function getUrl(): ?string;
     public function getPreparedUrl(): ?string;
@@ -30,80 +41,110 @@ interface Media extends Markup
     /**
      * @return static
      */
-    public function setId(?string $id): static;
+    public function setId(
+        ?string $id
+    ): static;
 
     public function getId(): ?string;
 
     /**
      * @return static
      */
-    public function setOrigin(?string $origin): static;
+    public function setOrigin(
+        ?string $origin
+    ): static;
+
     public function getOrigin(): ?string;
 
     /**
      * @return static
      */
-    public function setWidth(?int $width): static;
+    public function setWidth(
+        ?int $width
+    ): static;
 
     /**
      * @return static
      */
-    public function scaleWidth(int $width): static;
+    public function scaleWidth(
+        int $width
+    ): static;
 
     public function getWidth(): ?int;
 
     /**
      * @return static
      */
-    public function setHeight(?int $height): static;
+    public function setHeight(
+        ?int $height
+    ): static;
 
     public function getHeight(): ?int;
-    public function setDimensions(?int $width, ?int $height = null): static;
+
+    public function setDimensions(
+        ?int $width,
+        ?int $height = null
+    ): static;
 
     /**
      * @return static
      */
-    public function setAllowFullScreen(bool $flag): static;
+    public function setAllowFullScreen(
+        bool $flag
+    ): static;
+
     public function shouldAllowFullScreen(): bool;
 
     /**
      * @return static
      */
-    public function setAutoPlay(bool $flag): static;
+    public function setAutoPlay(
+        bool $flag
+    ): static;
 
     public function shouldAutoPlay(): bool;
 
     /**
      * @return static
      */
-    public function setStartTime(?int $seconds): static;
+    public function setStartTime(
+        ?int $seconds
+    ): static;
 
     public function getStartTime(): ?int;
 
     /**
      * @return static
      */
-    public function setEndTime(?int $seconds): static;
+    public function setEndTime(
+        ?int $seconds
+    ): static;
 
     public function getEndTime(): ?int;
 
     /**
      * @return static
      */
-    public function setDuration(?int $seconds): static;
+    public function setDuration(
+        ?int $seconds
+    ): static;
 
     public function getDuration(): ?int;
 
     /**
      * @param array<string, mixed> $options
      */
-    public function lookupThumbnail(?array $options = null): ?string;
+    public function lookupThumbnail(
+        ?array $options = null
+    ): ?string;
 
     /**
      * @param array<string, mixed> $options
      * @return array<string, mixed>|null
      */
-    public function lookupMeta(?array $options = null): ?array;
+    public function lookupMeta(
+        ?array $options = null
+    ): ?array;
 
     public function render(): ?Element;
     public function __toString(): string;