+++ src/Embed/MediaTrait.php
@@ -32,13 +32,15 @@ trait MediaTrait
     protected ?int $duration = null;
 
     protected ?string $source = null;
+    protected bool $consent = false;
 
 
     /**
      * Parse embed string
      */
-    public static function parse(string $embed): static
-    {
+    public static function parse(
+        string $embed
+    ): static {
         $embed = trim($embed);
         $stripEmbed = strip_tags($embed, '<iframe><object><embed><script>');
         $parts = explode('<', $stripEmbed, 2);
@@ -137,8 +139,9 @@ trait MediaTrait
     /**
      * Extract provider name from URL
      */
-    public static function extractProviderName(string $url): ?string
-    {
+    public static function extractProviderName(
+        string $url
+    ): ?string {
         foreach (self::URL_MAP as $search => $key) {
             if (false !== stripos($url, $search)) {
                 return $key;
@@ -151,8 +154,9 @@ trait MediaTrait
     /**
      * Get instance class for entry URL
      */
-    public static function getClassForUrl(string $url): string
-    {
+    public static function getClassForUrl(
+        string $url
+    ): string {
         $class = get_called_class();
 
         if ($provider = self::extractProviderName($url)) {
@@ -191,9 +195,12 @@ trait MediaTrait
 
     /**
      * Set media source URL
+     *
+     * @return $this
      */
-    protected function setUrl(?string $url): static
-    {
+    protected function setUrl(
+        ?string $url
+    ): static {
         if (empty($url)) {
             $this->url = null;
             return $this;
@@ -240,9 +247,12 @@ trait MediaTrait
 
     /**
      * Set Media element id
+     *
+     * @return $this
      */
-    public function setId(?string $id): static
-    {
+    public function setId(
+        ?string $id
+    ): static {
         $this->id = $id;
         return $this;
     }
@@ -257,9 +267,12 @@ trait MediaTrait
 
     /**
      * Set media origin
+     *
+     * @return $this
      */
-    public function setOrigin(?string $origin): static
-    {
+    public function setOrigin(
+        ?string $origin
+    ): static {
         $this->origin = $origin;
         return $this;
     }
@@ -275,18 +288,24 @@ trait MediaTrait
 
     /**
      * Set Media element width
+     *
+     * @return $this
      */
-    public function setWidth(?int $width): static
-    {
+    public function setWidth(
+        ?int $width
+    ): static {
         $this->width = $width;
         return $this;
     }
 
     /**
      * Scale dimensions from original width
+     *
+     * @return $this
      */
-    public function scaleWidth(int $width): static
-    {
+    public function scaleWidth(
+        int $width
+    ): static {
         $this->height = (int)round($width / $this->width * $this->height);
         $this->width = $width;
 
@@ -303,9 +322,12 @@ trait MediaTrait
 
     /**
      * Set Media element height
+     *
+     * @return $this
      */
-    public function setHeight(?int $height): static
-    {
+    public function setHeight(
+        ?int $height
+    ): static {
         $this->height = $height;
         return $this;
     }
@@ -320,6 +342,8 @@ trait MediaTrait
 
     /**
      * Set width and height and scale accordingly
+     *
+     * @return $this
      */
     public function setDimensions(
         ?int $width,
@@ -349,9 +373,12 @@ trait MediaTrait
 
     /**
      * Set whether Media element can go full screen
+     *
+     * @return $this
      */
-    public function setAllowFullScreen(bool $flag): static
-    {
+    public function setAllowFullScreen(
+        bool $flag
+    ): static {
         $this->allowFullScreen = $flag;
         return $this;
     }
@@ -367,9 +394,12 @@ trait MediaTrait
 
     /**
      * Set whether media can autoplay
+     *
+     * @return $this
      */
-    public function setAutoPlay(bool $autoplay): static
-    {
+    public function setAutoPlay(
+        bool $autoplay
+    ): static {
         $this->autoPlay = $autoplay;
         return $this;
     }
@@ -385,9 +415,12 @@ trait MediaTrait
 
     /**
      * Set start time of media
+     *
+     * @return $this
      */
-    public function setStartTime(?int $seconds): static
-    {
+    public function setStartTime(
+        ?int $seconds
+    ): static {
         $this->startTime = $seconds;
         return $this;
     }
@@ -402,9 +435,12 @@ trait MediaTrait
 
     /**
      * Set end time of media
+     *
+     * @return $this
      */
-    public function setEndTime(?int $seconds): static
-    {
+    public function setEndTime(
+        ?int $seconds
+    ): static {
         $this->endTime = $seconds;
 
         if ($this->endTime) {
@@ -424,9 +460,12 @@ trait MediaTrait
 
     /**
      * Set media duration
+     *
+     * @return $this
      */
-    public function setDuration(?int $seconds): static
-    {
+    public function setDuration(
+        ?int $seconds
+    ): static {
         $this->duration = $seconds;
 
         if ($this->duration) {
@@ -446,18 +485,41 @@ trait MediaTrait
 
 
     /**
+     * Set consent status
+     *
+     * @return $this
+     */
+    public function setConsent(
+        bool $consent
+    ): static {
+        $this->consent = $consent;
+        return $this;
+    }
+
+    /**
+     * Has consent
+     */
+    public function hasConsent(): bool
+    {
+        return $this->consent;
+    }
+
+
+    /**
      * Lookup thumbnail URL
      */
-    public function lookupThumbnail(?array $options = null): ?string
-    {
+    public function lookupThumbnail(
+        ?array $options = null
+    ): ?string {
         return null;
     }
 
     /**
      * Lookup media meta information
      */
-    public function lookupMeta(?array $options = null): ?array
-    {
+    public function lookupMeta(
+        ?array $options = null
+    ): ?array {
         return null;
     }