+++ src/Embed/Youtube.php
@@ -30,8 +30,9 @@ class Youtube extends Video
     /**
      * Extract parts from URL
      */
-    protected function setUrl(?string $url): static
-    {
+    protected function setUrl(
+        ?string $url
+    ): static {
         parent::setUrl($url);
 
         if (!$this->url) {
@@ -81,7 +82,13 @@ class Youtube extends Video
      */
     public function getPreparedUrl(): ?string
     {
-        $url = 'https://www.youtube.com/embed/' . $this->youtubeId;
+        $url = 'https://www.youtube';
+
+        if (!$this->consent) {
+            $url .= '-nocookie';
+        }
+
+        $url .= '.com/embed/' . $this->youtubeId;
         $queryVars = $this->options;
 
         if ($this->startTime !== null) {
@@ -135,16 +142,18 @@ class Youtube extends Video
     /**
      * Lookup thumbnail URL
      */
-    public function lookupThumbnail(?array $options = null): ?string
-    {
+    public function lookupThumbnail(
+        ?array $options = null
+    ): ?string {
         return 'https://img.youtube.com/vi/' . $this->youtubeId . '/hqdefault.jpg';
     }
 
     /**
      * Lookup media meta information
      */
-    public function lookupMeta(?array $options = null): ?array
-    {
+    public function lookupMeta(
+        ?array $options = null
+    ): ?array {
         $url = 'https://www.youtube.com/oembed?url=' . urlencode('https://www.youtube.com/watch?v=' . $this->youtubeId) . '&format=json';
         $infoUrl = 'https://www.youtube.com/get_video_info?video_id=' . $this->youtubeId;